New: compute scores of non-scoring rule ESCs. .xlsx files added
.xlsx files are collusion impact tables. How to read:
row = how much the country moves when removing the scores they handed out...
column = ... and the scores this country handed out
f.e. for collusion_1960_1980.xlsx, Ireland moved up 0.7333 places on average when removing the scores they handed out, and the scores Luxembourg handed out.
The entries that are filled in are based on Mantzaris et al. (2017)